[refactor] Create a file for the BasicErrorChecker in pylint.checker.base
